+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 CHANGELOG
 =========
 
+5.0.0 (2025-05-02)
+------------------
+* [FEATURE] The Opencast Course Overview is now accessible via the course navigation bar
+* [CHANGE] Most features from the Opencast Block plugin have been moved to
+the Opencast Tool plugin. The Opencast Block plugin is now optional.
+* [CHANGE] The course backup functionality has changed. The option to select
+individual events has been removed.\
+A site-wide admin setting (`importreducedduplication`) has been added.\
+If enabled, only events and series embedded via LTI or the Opencast
+activity module will be backed up.\
+If disabled (default), all events from
+the course will be included in the backup.
+
 4.5.3 (2025-01-16)
 ------------------
 * [FIX] #74 Refactor settings handling when propagating